Update vim-statuslines.md
The current GitBranch function will display a branch name for the current directory where Vim was launched even if a file is opened in a different directory outside of the git repository.  For example, when editing my ~/.vimrc the status line showed a git branch of "master" even though my .vimrc file in my home directory is not a part of a git repository.  This revision changes the command to cd to the open file's directory and detects the branch name from there.  

Also noted that command assumes the current file being edited is actually a committed member of the branch.