Add example function to repair corrupt ZSH history
Hello!

Thank you for this article. I ended up with a getting this exact error after my Pi hard crashed, and I had no idea what to do. Luckily, your article was the first thing that came up in Google. I really appreciate how easy you made it to understand the problem and the fix.

I'm just learning scripting and function writing in ZSH/Bash, and wanted to try to turn the script into a function, with added logic for locating the history file based on a defined $HISTFILE variable, which can be set as part of the standard ZSH setup on a new install. (It is set when a user runs ZSH for the first time after installing ZSH--or at least it is in the Arch Linux/Manjaro ARM version of ZSH.)

This is what I came up with and I wanted to share it in case someone might find it useful. I understand if it's not something you want to add to the article, though. :)

As I'm writing this comment, I realize that one potential issue is that the function might break if the user changes the value of $HISTFILE and does not reload/re-source the file containing the function. I don't know enough about ZSH/shell programming yet to know if there's a way to protect against that, beyond trusting the user not to do it.